refactoring to simplify code
